requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Data collection
requests>=2.31.0
beautifulsoup4>=4.12.0

# Configuration
pyyaml>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# CLI interface
click>=8.1.0
rich>=13.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
